Refactor match passing matrix class
The match passing matrix class has been refactored to divide the longest period of play into start and end times instead of a single endpoint. This allows for a more detailed analysis of players' playtime. Also, the pass events are added to the 'get' methods to retrieve pass information relating to each team. Sorting functionality has been added to ensure player on/off events are handled in chronological order.